requirements-llm.txt
# LLM and Interactive Game Dependencies
vllm>=0.6.0
openai>=1.0.0
httpx>=0.25.0

# For async support
aiofiles>=23.0.0

# Optional: For better terminal experience
rich>=13.0.0
click>=8.0.0

# Development and testing
pytest-asyncio>=0.21.0
